dRPB half fixes, enable FP16 for SM50, and more
This commit fixes an error in all half-precision RPB gradient kernels
which was introduced when GEMM kernels were integrated (see #38).
The rpb gradient pointer was incorrectly being reinterpret_cast-ed as a
float pointer, instead of a half pointer. This is likely because I was
planning to have these kernels store in FP32 as well as accumulate, but
ended up giving up on that without reverting the element types.
This issue was misattributed in #97.

We're also enabling FP16 for SM50 (no reason why it shouldn't work;
pending testing)

Other minor changes:
* Stop explicitly specifying torch releases in both requirement files
* Lint-related assertion in setup.py